+++ src/player.cpp
@@ -952,7 +952,8 @@ PlayerErrors Player::play(
     float pan,
     bool paused,
     bool looping,
-    double loopingStartAt)
+    double loopingStartAt,
+    double loopingEndAt)
 {
     ActiveSound *sound = findByHash(soundHash);
 
@@ -988,33 +989,39 @@ PlayerErrors Player::play(
 
     handle = 0;
     SoLoud::handle newHandle = 0;
+    const bool startPaused = paused || looping;
     if (busId == 0)
     {
-        newHandle = soloud.play(*sound->sound.get(), volume, pan, paused, 0);
+        newHandle = soloud.play(*sound->sound.get(), volume, pan, startPaused, 0);
     }
     else
     {
         auto it = busMap.find(busId);
         if (it != busMap.end())
-            newHandle = it->second.bus.play(*sound->sound.get(), volume, pan, paused);
+            newHandle = it->second.bus.play(*sound->sound.get(), volume, pan, startPaused);
         else
             return PlayerErrors::busIdNotFound;
+    }
+
+    if (newHandle != 0 && looping)
+    {
+        setLoopPoint(newHandle, loopingStartAt);
+        setLoopEndPoint(newHandle, loopingEndAt);
+        setLooping(newHandle, true);
     }
 
     if (newHandle != 0)
     {
         sound->handle.push_back({newHandle, MAX_DOUBLE, false});
+        if (looping)
+        {
+            setPause(newHandle, paused);
+        }
         // Check if this buffer has enough data to be played
         if (sound->soundType == SoundType::TYPE_BUFFER_STREAM)
         {
             static_cast<SoLoud::BufferStream *>(sound->sound.get())->checkBuffering(0);
         }
-    }
-
-    if (looping)
-    {
-        setLoopPoint(newHandle, loopingStartAt);
-        setLooping(newHandle, true);
     }
     handle = newHandle;
     return PlayerErrors::noError;
@@ -1204,6 +1211,16 @@ double Player::getLoopPoint(unsigned int handle)
 void Player::setLoopPoint(unsigned int handle, double time)
 {
     soloud.setLoopPoint(handle, time);
+}
+
+double Player::getLoopEndPoint(unsigned int handle)
+{
+    return soloud.getLoopEndPoint(handle);
+}
+
+void Player::setLoopEndPoint(unsigned int handle, double time)
+{
+    soloud.setLoopEndPoint(handle, time);
 }
 
 PlayerErrors Player::textToSpeech(const std::string &textToSpeech, unsigned int &handle)
@@ -1559,7 +1576,8 @@ PlayerErrors Player::play3d(
     bool paused,
     unsigned int busId,
     bool looping,
-    double loopingStartAt)
+    double loopingStartAt,
+    double loopingEndAt)
 {
     ActiveSound *sound = findByHash(soundHash);
     if (sound == 0)
@@ -1594,6 +1612,7 @@ PlayerErrors Player::play3d(
 
     handle = 0;
     SoLoud::handle newHandle = 0;
+    const bool startPaused = paused || looping;
     if (busId == 0)
     {
         newHandle = soloud.play3d(
@@ -1601,7 +1620,7 @@ PlayerErrors Player::play3d(
             posX, posY, posZ,
             velX, velY, velZ,
             volume,
-            paused,
+            startPaused,
             0);
     }
     else
@@ -1613,7 +1632,7 @@ PlayerErrors Player::play3d(
                 posX, posY, posZ,
                 velX, velY, velZ,
                 volume,
-                paused);
+                startPaused);
         else
             return PlayerErrors::busIdNotFound;
     }
@@ -1621,18 +1640,19 @@ PlayerErrors Player::play3d(
     if (newHandle != 0)
     {
         sound->handle.push_back({newHandle, MAX_DOUBLE, false});
+        if (looping)
+        {
+            setLoopPoint(newHandle, loopingStartAt);
+            setLoopEndPoint(newHandle, loopingEndAt);
+            setLooping(newHandle, true);
+            seek(newHandle, loopingStartAt);
+            setPause(newHandle, paused);
+        }
         // Check if this buffer has enough data to be played
         if (sound->soundType == SoundType::TYPE_BUFFER_STREAM)
         {
             static_cast<SoLoud::BufferStream *>(sound->sound.get())->checkBuffering(0);
         }
-    }
-    if (looping)
-    {
-        seek(newHandle, loopingStartAt);
-        setLoopPoint(newHandle, loopingStartAt);
-        setLooping(newHandle, true);
-        setPause(newHandle, paused);
     }
     handle = newHandle;
     return PlayerErrors::noError;