+++ src/soloud/include/soloud_audiosource.h
@@ -25,6 +25,8 @@ freely, subject to the following restrictions:
 #ifndef SOLOUD_AUDIOSOURCE_H
 #define SOLOUD_AUDIOSOURCE_H
 
+#include <stdint.h>
+
 #include "soloud.h"
 #include "soloud_fader.h"
 #include "soloud_filter.h"
@@ -187,6 +189,10 @@ namespace SoLoud
 		unsigned int mDelaySamples;
 		// When looping, start playing from this time
 		time mLoopPoint;
+		// When looping, wrap before this time. Zero uses the natural source end.
+		time mLoopEndPoint;
+		// Number of source frames decoded since the last seek.
+		uint64_t mSourceSamplePosition;
 
 		// Get N samples from the stream to the buffer. Report samples written.
 		virtual unsigned int getAudio(float *aBuffer, unsigned int aSamplesToRead, unsigned int aBufferSize) = 0;